refactor(backend): change paths to /blog/
