+++ apps/backend/src/app.module.ts
@@ -4,7 +4,7 @@ import { AppService } from './app.service'
 import { ConfigModule, ConfigService } from '@nestjs/config'
 import { MongooseModule } from '@nestjs/mongoose'
 import { SurveyModule } from './modules/survey/survey.module'
-import { BlogPostModule } from './modules/blog-post/blog-post.module'
+import { BlogPostModule } from './modules/blog/blog.module'
 
 @Module({
     imports: [
